Add saving message creation time to db and printing this for client
